+++ R/print.parameters_model.R
@@ -17,12 +17,15 @@
 #' @param show_formula Logical, if \code{TRUE}, adds the model formula to the output.
 #' @param caption Table caption as string. If \code{NULL}, no table caption is printed.
 #' @param footer_digits Number of decimal places for values in the footer summary.
-#' @param group Named character vector, can be used to group parameters in the
-#'   printed output. The lefthand-side (names) indicate the names of a group,
-#'   which will be inserted as "header row", while the righthand-side (values)
-#'   should match the name of a parameter where the group starts. A possible
-#'   use case might be to emphasize focal predictors and control variables,
-#'   see 'Examples'.
+#' @param groups Named list, can be used to group parameters in the printed output.
+#'   List elements may either be character vectors that match the name of those
+#'   parameters that belong to one group, or list elements can be row numbers
+#'   of those parameter rows that should belong to one group. The names of the
+#'   list elements will be used as group names, which will be inserted as "header
+#'   row". A possible use case might be to emphasize focal predictors and control
+#'   variables, see 'Examples'. Parameters will be re-ordered according to the
+#'   order used in \code{groups}, while all non-matching parameters will be added
+#'   to the end.
 #' @inheritParams insight::format_table
 #'
 #' @inheritSection format_parameters Interpretation of Interaction Terms
@@ -54,21 +57,39 @@
 #'   print(mp, select = "minimal")
 #' }
 #'
+#'
 #' # group parameters ------
 #'
 #' data(iris)
 #' model <- lm(
-#'   Sepal.Width ~ Species + Sepal.Length + Petal.Length,
+#'   Sepal.Width ~ Sepal.Length + Species + Petal.Length,
 #'   data = iris
 #' )
 #' # don't select "Intercept" parameter
 #' mp <- model_parameters(model, parameters = "^(?!\\(Intercept)")
-#' print(mp, group = c("Focal Predictors" = "Speciesversicolor",
-#'                     "Controls" = "Sepal.Length"))
+#' groups <- list(
+#'   "Focal Predictors" = c("Speciesversicolor", "Speciesvirginica"),
+#'   "Controls" = c("Sepal.Length", "Petal.Length")
+#' )
+#' print(mp, groups = groups)
 #'
-#' # only show coefficients, CI and p
-#' print(mp, group = c("Focal Predictors" = "Speciesversicolor",
-#'                     "Controls" = "Sepal.Length"), select = "minimal")
+#' # or use row indices
+#' print(mp, groups = list("Focal Predictors" = c(1, 4),
+#'                         "Controls" = c(2, 3)))
+#'
+#' # only show coefficients, CI and p,
+#' # put non-matched parameters to the end
+#'
+#' data(mtcars)
+#' mtcars$cyl <- as.factor(mtcars$cyl)
+#' mtcars$gear <- as.factor(mtcars$gear)
+#' model <- lm(mpg ~ hp + gear * vs + cyl + drat, data = mtcars)
+#'
+#' # don't select "Intercept" parameter
+#' mp <- model_parameters(model, parameters = "^(?!\\(Intercept)")
+
+#' print(mp, groups = list("Engine" = c("cyl6", "cyl8", "vs", "hp"),
+#'                         "Interactions" = c("gear4:vs", "gear5:vs")))
 #' }
 #' @export
 print.parameters_model <- function(x,
@@ -83,7 +104,7 @@ print.parameters_model <- function(x,
                                    show_sigma = FALSE,
                                    show_formula = FALSE,
                                    zap_small = FALSE,
-                                   group = NULL,
+                                   groups = NULL,
                                    ...) {
   # save original input
   orig_x <- x
@@ -120,7 +141,7 @@ print.parameters_model <- function(x,
     ci_width = "auto",
     ci_brackets = TRUE,
     format = "text",
-    group = group,
+    groups = groups,
     ...
   )