+++ R/print_md.R
@@ -18,7 +18,7 @@ print_md.parameters_model <- function(x,
                                       show_sigma = FALSE,
                                       show_formula = FALSE,
                                       zap_small = FALSE,
-                                      group = NULL,
+                                      groups = NULL,
                                       verbose = TRUE,
                                       ...) {
   # check if user supplied digits attributes
@@ -51,7 +51,7 @@ print_md.parameters_model <- function(x,
     ci_width = NULL,
     ci_brackets = ci_brackets,
     format = "markdown",
-    group = group,
+    groups = groups,
     ...
   )
 
@@ -94,9 +94,6 @@ print_md.compare_parameters <- function(x,
                                         p_digits = 3,
                                         style = NULL,
                                         ...) {
-  # save original input
-  orig_x <- x
-
   # check if user supplied digits attributes
   if (missing(digits)) {
     digits <- .additional_arguments(x, "digits", digits)